+++ scripts/package.js
@@ -25,8 +25,10 @@ const INCLUDED = [
   'popup.html',
   'popup.js',
   'help.html',
+  'help.js',
   'styles.css',
   'lib',
+  '_locales',
   'LICENSE',
   'README.md'
 ];